finalizando chamada para cards, iniciando cobertura de testes

+++ src/main/java/com/payment_service/interfaceadapters/presenters/dto/PaymentDto.java
@@ -28,7 +28,7 @@ public class PaymentDto {
 
     private String paymentDescription;
 
-    private BigDecimal paymentValue;
+    private double paymentValue;
 
     private LocalDateTime transactionDate;
